backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pyotp==2.9.0
cryptography==41.0.7
gunicorn==21.2.0
pytest==7.4.3
pytest-mock==3.12.0
hypothesis==6.92.1

# ML Libraries
torch==2.9.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
tensorflow==2.18.0

# WebSocket Support
Flask-SocketIO==5.3.6
python-socketio[client]==5.11.2
eventlet==0.35.2

# Redis for Caching
redis==5.0.1
flask-caching==2.1.0

# Message Queue (Celery + RabbitMQ)
celery==5.3.4
kombu==5.3.4

# Blockchain Integration
web3==6.11.3
py-solc-x==2.0.2

# IPFS Integration
ipfshttpclient==0.8.0a2

# AI Integration (Claude API)
anthropic>=0.74.0

# Geolocation and IP Reputation
geoip2==4.7.0
requests==2.31.0

# NLP Libraries
nltk==3.8.1
spacy==3.7.2

# Additional Utilities
geopy==2.4.1
python-dateutil==2.8.2
psutil==5.9.6

# Enhanced Zero Trust Dependencies
pydantic==2.5.3
Pillow==10.1.0
qrcode[pil]==7.4.2

# OAuth 2.0 and API Security
authlib==1.3.0
flask-limiter==3.5.0
marshmallow==3.20.2
apispec==6.3.0
apispec-webframeworks==0.5.2

# Security and Input Sanitization
bleach==6.1.0
validators==0.22.0

# Monitoring and Observability
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
sentry-sdk[flask]==1.39.2
structlog==23.2.0
elasticsearch==8.11.1
logstash-logger==0.1.0

# Disaster Recovery and Backup Dependencies
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0
schedule==1.2.0
